fenlei

+++ src/main/java/top/zhaogaoshang/store/controller/Category.java
@@ -6,6 +6,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 import top.zhaogaoshang.store.common.Utils;
 import top.zhaogaoshang.store.mapper.OneCategoryMapper;
+import top.zhaogaoshang.store.mapper.ProductMapper;
 import top.zhaogaoshang.store.mapper.TwoCategoryMapper;
 import top.zhaogaoshang.store.model.OneCategory;
 import top.zhaogaoshang.store.model.Paging;
@@ -15,6 +16,8 @@ import top.zhaogaoshang.store.model.TwoCategory;
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Iterator;
 
 /**
  * 赵高尚
@@ -32,9 +35,6 @@ public class Category {
     TwoCategoryMapper twoCategoryMapper;
 
     @Resource
-    TwoCategory twoCategory;
-
-    @Resource
     Result result;
 
     @Resource
@@ -44,7 +44,35 @@ public class Category {
     OneCategory oneCategory;
 
     @Resource
+    TwoCategory twoCategory;
+
+    @Resource
     Paging paging;
+
+    @Resource
+    ProductMapper productMapper;
+
+    /*
+     * 更改分类
+     * */
+    @PostMapping("/edit")
+    public Object putCategory(HttpServletRequest httpServletRequest) {
+        Object type =  httpServletRequest.getParameter("type");
+        HashMap map = new HashMap();
+        System.out.println("获取的结果为 " + type);
+
+        // 一级分类
+        if ("1".equals(type)) {
+            map.put("oneCategoryId",httpServletRequest.getParameter("oneCategoryId"));
+            map.put("oneCategoryName",httpServletRequest.getParameter("editCategoryParamsName"));
+            return this.updataOneCategory(map);
+        } else {
+            // 二级分类
+            map.put("twoCategoryId",httpServletRequest.getParameter("twoCategoryId"));
+            map.put("twoCategoryName",httpServletRequest.getParameter("editCategoryParamsName"));
+            return this.twoCategoryUpdata(map);
+        }
+    }
 
     /*
      * 一级分类
@@ -66,10 +94,17 @@ public class Category {
      * 删除一级分类
      * */
     @DeleteMapping("/one/delete")
-    public Object deleteOneCategory(Integer userType, String categoryId) {
+    public Object deleteOneCategory(String categoryId) {
         // 权限
-        if (userType != 1) {
-            result.error("权限不足");
+//        if (userType != 1) {
+//            result.error("您权限不足,请使用仓库创建");
+//            return result;
+//        }
+
+        // 一级分类下是不是有二级分类
+        ArrayList<TwoCategory> twoCategories = twoCategoryMapper.selectByOneCategoryId(categoryId, 1); // 删除一级分类
+        if (!twoCategories.isEmpty()) {
+            result.error("在此分类下有二级分类，请先删除二级分类");
             return result;
         }
 
@@ -85,17 +120,17 @@ public class Category {
     /*
      * 更改一级分类名字
      * */
-    @PutMapping("/one/updata")
-    public Object updataOneCategory(HttpServletRequest httpServletRequest) {
+    @PostMapping("/one/updata")
+    public Object updataOneCategory(HashMap map) {
         // 权限
-        String userType = httpServletRequest.getParameter("userType");
-        if (!"1".equals(userType)) {
-            result.error("权限不够");
-            return result;
-        }
+//        String userType = httpServletRequest.getParameter("userType");
+//        if (!"1".equals(userType)) {
+//            result.error("权限不够");
+//            return result;
+//        }
 
-        oneCategory.setOneCategoryId(httpServletRequest.getParameter("oneCategoryId"));
-        oneCategory.setOneCategoryName(httpServletRequest.getParameter("oneCategoryName"));
+        oneCategory.setOneCategoryId((String) map.get("oneCategoryId"));
+        oneCategory.setOneCategoryName((String) map.get("editCategoryParamsName"));
         int i = oneCategoryMapper.updateByPrimaryKeySelective(oneCategory);
 
         if (i == 1) {
@@ -107,9 +142,6 @@ public class Category {
         return result;
     }
 
-    /*
-     * 查找一级分类详情
-     * */
     /*
      * 查找一级分类列表
      * */
@@ -145,6 +177,14 @@ public class Category {
      * */
     @DeleteMapping("/two/delete")
     public Object twoCategoreDelete(TwoCategory twoCategory) {
+
+        Integer integer = productMapper.selectAllByCategory(twoCategory.getTwoCategoryId()); // 是否有商品
+        System.out.println("二级分类下的商品数量：" + integer);
+        if (integer > 0) {
+            result.error("二级分类下有商品，请先清空商品");
+            return result;
+        }
+
         int i = twoCategoryMapper.deleteByPrimaryId(twoCategory.getTwoCategoryId());
         if (i == 1) {
             result.success("删除成功");
@@ -158,7 +198,11 @@ public class Category {
      * 更改二级分类
      * */
     @PutMapping("two/updata")
-    public Object twoCategoryUpdata(TwoCategory twoCategory) {
+    public Object twoCategoryUpdata(HashMap map) {
+
+        twoCategory.setTwoCategoryId((String) map.get("twoCategoryId"));
+        twoCategory.setTwoCategoryName((String) map.get("editCategoryParamsName"));
+
         int i = twoCategoryMapper.updateByPrimaryIdSelective(twoCategory);
         if (i == 1) {
             result.success("更改成功");
@@ -172,9 +216,12 @@ public class Category {
      * 根据一级分类查找二级分类
      * */
     @GetMapping("/two/by/category/list")
-    public Object twoCategoryByList(String oneCategoryId) {
-        ArrayList<TwoCategory> twoCategories = twoCategoryMapper.selectByOneCategoryId(oneCategoryId);
-        result.success(twoCategories);
+    public Object twoCategoryByList(String oneCategoryId, Integer page) {
+        Integer integer = twoCategoryMapper.selectByAllTwoCategory(oneCategoryId); // 所有二级分类
+        System.out.println(integer);
+        ArrayList<TwoCategory> twoCategories = twoCategoryMapper.selectByOneCategoryId(oneCategoryId, page);
+        paging.result(integer, twoCategories, page);
+        result.success(paging);
         return result;
     }
